feat: 불필요한 모듈 제거 및 핵심 모델/서비스 로직 리팩토링
- 제거: Part, Material, Assembly 관련 모델, 스키마, 라우터 파일 삭제
- 수정: Lot, Pallet, Process, RFID 모델 정의 업데이트
- 수정: RFID 처리 및 Trace 서비스 로직 개선
- 정리: __pycache__ 및 불필요한 파일 정리

+++ implementation/api/app/models/lot.py
@@ -15,7 +15,7 @@ class Lot(BaseModel):
     initial_quantity = Column(Integer, nullable=False, comment='초기 수량')
     status = Column(String(20), default='WAIT', comment='LOT 상태 (WAIT, PROCESS, STOCK, CONSUMED, SHIPPED, HOLD, DEFECT)')
     production_date = Column(Date, nullable=False, comment='생산일 또는 입고일')
-    process_id = Column(BigInteger, ForeignKey("processes.id"), comment='생성된 공정 ID')
+    process_id = Column(Integer, ForeignKey("processes.id"), comment='생성된 공정 ID')
     supplier = Column(String(100), comment='공급사 (원자재 입고 시, 기본 공급사와 다를 경우)')
     worker_name = Column(String(50), comment='작업자')
     qc_passed = Column(Boolean, default=False, comment='QC 합격 여부')
